ingest into postgres

+++ cli/cmd/ingestion/neurons.go
@@ -1,32 +1,29 @@
 package ingestion
 
 import (
-	"database/sql"
 	"errors"
 	"github.com/charmbracelet/log"
 )
 
 type Neuron struct {
-	id                 int
-	uid                string
-	embryonic          bool
-	filename           string
-	fileHash           string
-	developmentalStage sql.NullInt64
-	timepoint          int
+	id        int
+	uid       string
+	embryonic bool
+	filename  string
+	timepoint int
 }
 
 // GetNeuron gets the neuron by UID and returns it, taking an optional timepoint and developmental stage
 // the timepoint defaults to 0, and the developmental stage defaults to 0
-func (n *Neuroscan) GetNeuron(uid string, timepoint int, devStage sql.NullInt64) (Neuron, error) {
-	log.Debug("Getting neuron", "uid", uid, "timepoint", timepoint, "devStage", devStage)
+func (n *Neuroscan) GetNeuron(uid string, timepoint int) (Neuron, error) {
+	log.Debug("Getting neuron", "uid", uid, "timepoint", timepoint)
 	var neuron Neuron
 
-	query := "SELECT neurons.id, uid, embryonic, filename, file_hash, developmental_stage, timepoint FROM neurons WHERE uid = $1 AND timepoint = $2 AND developmental_stage = $3"
+	query := "SELECT neurons.id, uid, embryonic, filename, timepoint FROM neurons WHERE uid = $1 AND timepoint = $2"
 
-	args := []interface{}{uid, timepoint, devStage}
+	args := []interface{}{uid, timepoint}
 
-	err := n.connPool.QueryRow(n.context, query, args...).Scan(&neuron.id, &neuron.uid, &neuron.embryonic, &neuron.filename, &neuron.fileHash, &neuron.developmentalStage, &neuron.timepoint)
+	err := n.connPool.QueryRow(n.context, query, args...).Scan(&neuron.id, &neuron.uid, &neuron.embryonic, &neuron.filename, &neuron.timepoint)
 	if err != nil {
 		log.Error("Error getting neuron: ", "uid", uid, "timepoint", timepoint, "err", err)
 		return neuron, err
@@ -57,7 +54,7 @@ func (neuron Neuron) writeToDB(n *Neuroscan) {
 		}
 	}
 
-	err = n.CreateNeuron(neuron.uid, neuron.embryonic, neuron.filename, neuron.fileHash, neuron.developmentalStage, neuron.timepoint)
+	err = n.CreateNeuron(neuron.uid, neuron.embryonic, neuron.filename, neuron.timepoint)
 	if err != nil {
 		log.Error("Error inserting new neuron record", "err", err)
 		return
@@ -153,6 +150,10 @@ func (n *Neuroscan) NeuronExists(uid string, timepoint int) (bool, error) {
 
 // DeleteNeuron deletes a neuron record from the database by the given uid and timepoint
 func (n *Neuroscan) DeleteNeuron(uid string, timepoint int) error {
+	if n.connPool == nil {
+		return errors.New("connection pool is nil")
+	}
+
 	_, err := n.connPool.Exec(n.context, "DELETE FROM neurons WHERE uid = $1 AND timepoint = $2", uid, timepoint)
 	if err != nil {
 		return err
@@ -162,7 +163,7 @@ func (n *Neuroscan) DeleteNeuron(uid string, timepoint int) error {
 }
 
 // CreateNeuron creates a new neuron record in the database
-func (n *Neuroscan) CreateNeuron(uid string, embryonic bool, filename string, fileHash string, developmentalStage sql.NullInt64, timepoint int) error {
+func (n *Neuroscan) CreateNeuron(uid string, embryonic bool, filename string, timepoint int) error {
 	exists, err := n.NeuronExists(uid, timepoint)
 
 	if err != nil {
@@ -173,18 +174,7 @@ func (n *Neuroscan) CreateNeuron(uid string, embryonic bool, filename string, fi
 		return errors.New("neuron already exists")
 	}
 
-	_, err = n.connPool.Exec(n.context, "INSERT INTO neurons (uid, embryonic, filename, file_hash, timepoint, developmental_stage) VALUES ($1, $2, $3, $4, $5, $6)", uid, embryonic, filename, fileHash, timepoint, developmentalStage)
-	if err != nil {
-		return err
-	}
-
-	newNeuron, err := n.GetNeuron(uid, timepoint, developmentalStage)
-
-	if err != nil {
-		return err
-	}
-
-	_, err = n.connPool.Exec(n.context, "INSERT INTO neurons__developmental_stages (neuron_id, \"developmental-stage_id\") VALUES ($1, $2)", newNeuron.id, developmentalStage)
+	_, err = n.connPool.Exec(n.context, "INSERT INTO neurons (uid, embryonic, filename, timepoint) VALUES ($1, $2, $3, $4)", uid, embryonic, filename, timepoint)
 	if err != nil {
 		return err
 	}
@@ -203,20 +193,11 @@ func parseNeuron(n *Neuroscan, filePath string) (Neuron, error) {
 
 	fileMeta := fileMetas[0]
 
-	devStage, err := n.GetDevStageByUID(fileMeta.developmentalStage)
-
-	if err != nil {
-		log.Error("Error getting dev stage ID", "err", err)
-		return Neuron{}, err
-	}
-
 	neuron := Neuron{
-		uid:                fileMeta.uid,
-		embryonic:          false,
-		filename:           fileMeta.filename,
-		fileHash:           fileMeta.filehash,
-		developmentalStage: devStage.id,
-		timepoint:          fileMeta.timepoint,
+		uid:       fileMeta.uid,
+		embryonic: false,
+		filename:  fileMeta.filename,
+		timepoint: fileMeta.timepoint,
 	}
 
 	return neuron, nil
